feat: add categorical class-probability forecasting (#30)
* feat: add categorical class-probability forecasting


* fix: audit cleanup for class-proba forecasting

- Add n_classes_ fitted attribute to BaseClassProbaForecaster and
  ClassProbaReductionForecaster (matches plan requirement)
- Add predict_class_proba/observe_predict_class_proba to SearchCV
  metadata routing
- Add numeric type validation for probability columns in
  validate_scorer_data
- Update check_class_proba_classes_attribute to verify n_classes_

* feat: add classification datasets, plotting support, and example notebooks

- Add fetch_air_quality_classification and fetch_demand_classification datasets
- Remove make_weather_classification generator in favor of real datasets
- Add class probability calibration plot to evaluation module
- Add categorical/class_proba detection and plotting to forecasting module
- Remove standalone plot_class_probabilities (merged into plot_forecast)
- Move class_proba example from point/ to dedicated class_proba/ directory
- Add class_proba_metrics example notebook
- Update existing examples for class-proba forecasting support
- Fix predict_class_proba routing order in search module
- Add tests for new datasets and plotting functions


* chore: remove mypy type: ignore comments and fix ty errors

- Remove all 54 `# type: ignore` comments (mypy not used in this project)
- Add `# ty: ignore` suppressions for ty 0.0.29 type checker errors
- Use `# fmt: skip` where ruff formatting would split ty ignores across lines
- Add module docstring to empty _generators.py (interrogate 100% coverage)
- Fix corrupted ignore comment in tests/strategies.py
- Add `noqa: PLC0415` for intentional lazy import in _compat.py

* fix(ci): fix docstring examples, example notebooks, and docs build

- Fix BrierScore and LogLoss docstring expected values
- Fix observe_predict_workflow notebook: use observe+predict pattern
  instead of observe_predict_* for plotting (avoids unsorted time)
- Fix class_proba_metrics notebook: avoid double-observe on same
  forecaster by using observe+predict_class_proba pattern
- Fix forecasting_visualization notebook: correct target name
  (air_quality, not quality)
- Fix docs build: apply _skip_stems check for nixtla notebooks
  when yohou_nixtla is not installed

* fix(examples): use observe_predict_* methods directly in notebooks

- Use observe_predict_interval() with .tail(fh) in workflow notebook
- Use observe_predict_class_proba() with .tail(fh) for soft classification
- Use observe_predict() with .tail(fh) for hard-label classification
- Use deepcopy to create independent fitted copies when two cells
  need to call observe_predict_* on the same forecaster type
- Restore original section titles and markdown descriptions

* refactor(examples): remove nixtla notebooks and simplify observe_predict usage

- Remove nixtla_forecasters.py and nixtla_panel.py (moved to yohou-nixtla)
- Remove nixtla skip logic from docs/hooks.py and test marks from test_examples.py
- Drop unnecessary .tail(fh) from observe_predict_* calls in notebooks
- Drop unnecessary .head(fh) from y_test in plot_forecast calls
- Keep .sort('time') since check_time_column requires ascending order

* test: improve patch coverage for class-probability forecasting

- Add time_weight tests (global and panel) for LogLoss, BrierScore, Accuracy
- Add componentwise aggregation rename tests for all three scorers
- Add unknown label fallback test for LogLoss
- Add class_proba panel/component tests for check_scorer_column_selection
- Add predict_class_proba/observe_predict_class_proba tests for ColumnForecaster
- Add predict_class_proba/observe_predict_class_proba tests for ForecastedFeatureForecaster
- Add recursive prediction (horizon > fit) tests for ClassProbaReductionForecaster
- Add panel data tests for direct and multi-output strategies
- Add systematic panel checks with expected failures for known issues
- Add GridSearchCV class_proba predict/observe_predict test
- Fix panel data bug: use unprefixed column names in classes_/label_to_code_

* fix(ci): fix lint formatting and add patch coverage tests

- Fix ruff format issues across 12 source files
- Add PLC0415 to per-file-ignores for docs/**/*
- Add coverage tests for class_proba code paths:
  - validate_data scorer validation (missing/non-numeric proba columns)
  - model_selection forecaster-scorer compatibility checks
  - polars cast non-numeric type branch
  - scorer prediction type compatibility check
  - ColumnForecaster predict_class_proba with remainder
  - ForecastedFeatureForecaster predict_class_proba with X
  - plot_forecast panel class_proba with training data

* fix(ci): fix ty type errors and add patch coverage tests

- Fix 3 unsupported-operator ty errors in diagnostics.py by casting
  polars .sum() results to float before division
- Add coverage tests:
  - FeaturePipeline.n_features_in_ and feature_names_in_ properties
  - Single-target componentwise aggregation for LogLoss/BrierScore/Accuracy
  - Multi-target class_proba reduction with multi-output/direct strategies
  - Multi-step interval prediction with panel data
  - Multi-step interval prediction with point/mean strategies

---------

+++ src/yohou/plotting/diagnostics.py
@@ -138,8 +138,8 @@ def plot_autocorrelation(
                 else:
                     s1 = series[:-lag] - mean_v
                     s2 = series[lag:] - mean_v
-                    num = (s1 * s2).sum()
-                    den = ((series - mean_v) ** 2).sum()
+                    num = float((s1 * s2).sum())
+                    den = float(((series - mean_v) ** 2).sum())
                     acf_vals.append(num / den if den != 0 else 0)
             fig.add_trace(
                 go.Bar(x=list(range(_ml + 1)), y=acf_vals, marker={"color": _bc}, showlegend=False),
@@ -206,8 +206,8 @@ def plot_autocorrelation(
                 # Compute correlation at lag
                 series1 = series[:-lag] - mean_val
                 series2 = series[lag:] - mean_val
-                numerator = (series1 * series2).sum()
-                denominator = ((series - mean_val) ** 2).sum()
+                numerator = float((series1 * series2).sum())
+                denominator = float(((series - mean_val) ** 2).sum())
                 acf = numerator / denominator if denominator != 0 else 0
                 acf_values.append(acf)
 
@@ -370,7 +370,7 @@ def _compute_pacf(
         return _compute_pacf_durbin_levinson(values, nlags, alpha)
 
     if alpha is not None:
-        pacf_result, confint = sm_pacf(values, nlags=nlags, method=method, alpha=alpha)  # type: ignore[arg-type]
+        pacf_result, confint = sm_pacf(values, nlags=nlags, method=method, alpha=alpha)  # fmt: skip  # ty: ignore[invalid-argument-type]
         # statsmodels returns CIs centered on the PACF values (pacf ± z/√n).
         # For significance testing we need horizontal bands at ±z/√n (centered
         # on zero), so subtract the PACF values from both bounds.
@@ -378,8 +378,8 @@ def _compute_pacf(
         ci_hi = (confint[:, 1] - pacf_result).tolist()
         return pacf_result.tolist(), ci_lo, ci_hi
 
-    pacf_result = sm_pacf(values, nlags=nlags, method=method)  # type: ignore[arg-type]
-    return pacf_result.tolist(), None, None  # type: ignore[union-attr]
+    pacf_result = sm_pacf(values, nlags=nlags, method=method)  # ty: ignore[invalid-argument-type]
+    return pacf_result.tolist(), None, None  # ty: ignore[unresolved-attribute]
 
 
 def plot_partial_autocorrelation(
@@ -1495,8 +1495,8 @@ def plot_lag_scatter(
                 if show_diagonal:
                     source = df_aug if df_aug is not None else df
                     dl_all = source.with_columns(pl.col(col).shift(lag).alias("lagged")).drop_nulls()
-                    vmin = min(float(dl_all[col].min()), float(dl_all["lagged"].min()))  # type: ignore[arg-type]
-                    vmax = max(float(dl_all[col].max()), float(dl_all["lagged"].max()))  # type: ignore[arg-type]
+                    vmin = min(float(dl_all[col].min()), float(dl_all["lagged"].min()))  # fmt: skip  # ty: ignore[invalid-argument-type]
+                    vmax = max(float(dl_all[col].max()), float(dl_all["lagged"].max()))  # fmt: skip  # ty: ignore[invalid-argument-type]
                     fig.add_trace(
                         go.Scatter(
                             x=[vmin, vmax],
@@ -1607,11 +1607,11 @@ def plot_lag_scatter(
             x_mean = y_lagged.mean()
             y_mean = y_current.mean()
 
-            numerator = ((y_lagged - x_mean) * (y_current - y_mean)).sum()
-            denominator = ((y_lagged - x_mean) ** 2).sum()
+            numerator = float(((y_lagged - x_mean) * (y_current - y_mean)).sum())
+            denominator = float(((y_lagged - x_mean) ** 2).sum())
 
             if denominator != 0 and x_mean is not None and y_mean is not None:
-                slope = float(numerator / denominator)
+                slope = numerator / denominator
                 intercept = cast(float, y_mean) - slope * cast(float, x_mean)
 
                 y_lagged_min = y_lagged.min()
@@ -2010,7 +2010,7 @@ def plot_scatter_matrix(
                             pass
                         else:
                             data_range = float(vals.max()) - float(vals.min())
-                            if data_range > 0 and kde.factor > 1e-10:
+                            if data_range > 0 and kde.factor > 1e-10:  # ty: ignore[unsupported-operator]
                                 x_grid = np.linspace(float(vals.min()), float(vals.max()), 200)
                                 fig.add_trace(
                                     go.Scatter(
